fix IIS_DownloadZipToWebsite

+++ vars/IIS_DownloadZipToWebsite.groovy
@@ -1,6 +1,9 @@
 #!/usr/bin/env groovy
 def call() {
     textWithColor("IIS_DownloadZipToWebsite")
+    if(isEmpty(env.IIS_FOLDER)) {
+        throw new Exception('env.IIS_FOLDER is must')
+    }
     bashCommand("rm -rf ./Artifact.zip || true")
     downloadFile('./Artifact.zip',env.ArtifactUrl)
     bashCommand("ls -latr")
